Dataset Pane refactor
