+++ matico_components/src/Components/MaticoEditor/Panes/DatasetsEditor.tsx
@@ -12,7 +12,8 @@ import {
     TabList,
     Item,
     TabPanels,
-    StatusLight
+    StatusLight,
+    Divider
 } from "@adobe/react-spectrum";
 import React, { useState } from "react";
 import { DatasetModal } from "./DatasetModal";
@@ -21,18 +22,19 @@ import { useMaticoDispatch, useMaticoSelector } from "Hooks/redux";
 import { addDataset } from "Stores/MaticoSpecSlice";
 import { DatasetProvider } from "Datasets/DatasetProvider";
 import { Dataset as DatasetSpec } from "@maticoapp/matico_types/spec";
+import {DatasetTransfromPane} from "../EditorComponents/DatasetTransformPane/DatasetTransformPane";
 
 interface DatasetEditorProps {
     dataset: DatasetSummary;
 }
+export const DatasetStatusColors: Record<DatasetState, "yellow" | "negative" | "positive"> =
+    {
+        [DatasetState.LOADING]: "yellow",
+        [DatasetState.ERROR]: "negative",
+        [DatasetState.READY]: "positive"
+    };
 
 export const DatasetEditor: React.FC<DatasetEditorProps> = ({ dataset }) => {
-    let statusColors: Record<DatasetState, "yellow" | "negative" | "positive"> =
-        {
-            [DatasetState.LOADING]: "yellow",
-            [DatasetState.ERROR]: "negative",
-            [DatasetState.READY]: "positive"
-        };
 
     return (
         <DatasetModal dataset={dataset}>
@@ -44,7 +46,7 @@ export const DatasetEditor: React.FC<DatasetEditorProps> = ({ dataset }) => {
                 justifyContent="space-between"
             >
                 <Text>{dataset.name}</Text>
-                <StatusLight variant={statusColors[dataset.state]} />
+                <StatusLight variant={DatasetStatusColors[dataset.state]} />
             </Flex>
         </DatasetModal>
     );
@@ -61,7 +63,7 @@ const NewDatasetModal: React.FC<NewDatasetModalProps> = ({
 }) => {
     return (
         <DialogTrigger isDismissable>
-            <ActionButton>Add</ActionButton>
+            <ActionButton >Add Dataset</ActionButton>
             {(close) => (
                 <Dialog>
                     <Content>
@@ -121,25 +123,35 @@ export const DatasetsEditor: React.FC<DatasetsEditorProps> = ({
 
     return (
         <Flex height="100%" width="100%" direction="column">
-            <Heading margin="size-150" alignSelf="start">
-                Datasets
-            </Heading>
-            <Flex direction="column">
-                {Object.entries(datasets).map(([datasetName, dataset]) => {
-                    return (
-                        <ActionButton width="100%">
+            <Flex direction='column' gap="size-500">
+              <Heading margin="size-150" alignSelf="start">
+                  Datasets
+              </Heading>
+              <Flex direction="column">
+                <>
+                {Object.entries(datasets).filter(([datasetName,dataset])=>!dataset.transform).map(([datasetName, dataset]) => {
+                      return (
                             <DatasetEditor
                                 dataset={dataset}
                                 key={datasetName}
                             />
-                        </ActionButton>
-                    );
-                })}
+                      );
+                  })}
+                </>
+
+              <NewDatasetModal
+                  onSubmit={createDataset}
+                  datasetProviders={datasetProviders}
+              />
+              <Well marginTop="size-1000" alignSelf="end">
+                Load Datasets from a variety of sources 
+              </Well>
+              </Flex>
             </Flex>
-            <NewDatasetModal
-                onSubmit={createDataset}
-                datasetProviders={datasetProviders}
-            />
+            <Divider size="S"/>
+            <Flex direction="column">
+                  <DatasetTransfromPane />
+            </Flex>
         </Flex>
     );
 };
